yolo5/requirements.txt
pyyaml
loguru
boto3
requests>=2.31.0